refactor: turn into util function
